refactor calculators - remove calculatorReducer, replace with custom state and effects
